scalesafe/backend/requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
psutil==5.8.0
boto3==1.18.57
google-api-python-client==2.70.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
